gram2vec feature highlights; restructuring utils for clarity
